refactor :: id 받기
