+++ src/main/java/com/example/daedongv3_5/domain/auth/presentation/AuthController.java
@@ -2,18 +2,13 @@ package com.example.daedongv3_5.domain.auth.presentation;
 
 import com.example.daedongv3_5.domain.auth.presentation.dto.request.LoginRequest;
 import com.example.daedongv3_5.domain.auth.presentation.dto.response.TokenResponse;
+import com.example.daedongv3_5.domain.auth.service.NotifyRecruitmentResultService;
 import com.example.daedongv3_5.domain.auth.service.ReissueTokenService;
 import com.example.daedongv3_5.domain.auth.service.StudentLoginService;
 import com.example.daedongv3_5.domain.auth.service.TeacherLoginService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
@@ -23,6 +18,7 @@ public class AuthController {
     private final ReissueTokenService reissueTokenService;
     private final StudentLoginService studentLoginService;
     private final TeacherLoginService teacherLoginService;
+    private final NotifyRecruitmentResultService notifyRecruitmentResultService;
 
     @PostMapping("/student")
     @ResponseStatus(HttpStatus.OK)
@@ -40,5 +36,15 @@ public class AuthController {
     @ResponseStatus(HttpStatus.OK)
     TokenResponse reissueToken(@RequestHeader(name = "x-refresh-token") String token) {
         return reissueTokenService.reissueToken(token);
+    }
+
+    @PostMapping("/pass/{id}")
+    public void pass(@PathVariable Long id) {
+        notifyRecruitmentResultService.pass(id);
+    }
+
+    @PostMapping("/failed/{id}")
+        public void failed(@PathVariable Long id) {
+            notifyRecruitmentResultService.failed(id);
     }
 }
